Converter library refactor stage 14
Moved the converter implementation files out of src/export and into format-owned directories:

  - src/formats/enigmaxml
  - src/formats/mnx
  - src/formats/mss
  - src/formats/svg

  src/export is now left as CLI export-command plumbing, with includes/CMake/test paths updated to the new layout. I also checked for stale export/... format includes and found
  none.
