+++ src/formats/enigmaxml/enigmaxml.cpp
@@ -41,7 +41,7 @@
 #include "musx/musx.h"
 
 #include "denigma.h"
-#include "export/enigmaxml.h"
+#include "enigmaxml.h"
 #include "utils/ziputils.h"
 #include "score_encoder/score_encoder.h"
