+++ src/formats/mnx/mnx_enums.cpp
