+++ src/formats/mnx/mnx_formatted_text.cpp
