+++ src/formats/mnx/mnx_mapping.cpp
