+++ src/formats/mnx/mnx_mapping.h
