+++ src/formats/mss/mss.h
