+++ src/formats/svg/svg.h
