+++ tests/mnx/test_global.cpp
@@ -29,7 +29,7 @@
 #include "mnxdom.h"
 #include "test_utils.h"
 #include "musx/musx.h"
-#include "export/mnx.h"
+#include "formats/mnx/mnx.h"
 
 using namespace denigma;
 using namespace musx::dom;
